requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.1
Flask==2.1.2
Flask-Admin==1.6.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
Werkzeug==2.1.2
WTForms==3.0.1
